backend/requirements.txt
PyMuPDF==1.25.3
Flask==3.1.0
flask-cors==5.0.1
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
supabase>=2.0.0
